OAK-10118: (wip) reverted test with includePropertyNames/orderedProps use
