+++ oak-search-elastic/src/test/java/org/apache/jackrabbit/oak/plugins/index/elastic/ElasticOrderByCommonTest.java
@@ -3,8 +3,6 @@ package org.apache.jackrabbit.oak.plugins.index.elastic;
 import org.apache.jackrabbit.oak.api.ContentRepository;
 import org.apache.jackrabbit.oak.plugins.index.OrderByCommonTest;
 import org.junit.ClassRule;
-import org.junit.Ignore;
-import org.junit.Test;
 
 public class ElasticOrderByCommonTest extends OrderByCommonTest {
 
@@ -27,10 +25,4 @@ public class ElasticOrderByCommonTest extends OrderByCommonTest {
         setTraversalEnabled(false);
     }
 
-    @Test
-    @Override
-    @Ignore("No support for includePropertyNames/orderedProps in elastic")
-    public void sortQueriesWithLong_NotIndexed() throws Exception {
-        super.sortQueriesWithLong_NotIndexed();
-    }
 }
